chore: Adapt tests to the changes in the code
